AE-2182 AED-741 connected dev tests to prod
