+++ defender/tests/datadog_alerts_heat_map.spec.js
@@ -1,5 +1,5 @@
 // Set the name of the Secret set in Autotask
-const stackName = 'datadog_forta_bot_alerts_dev';
+const stackName = 'datadog_alerts_heat_map';
 const datadogSecretName = `${stackName}_datadogApiKey`;
 
 jest.mock('axios', () => jest.fn());
@@ -15,24 +15,22 @@ const mockKeyValueStore = {
 };
 
 const botIdsToNames = {
-  '0x5a00b44b2db933d4c797e6bd3049abdeb89cc9ec1b2eaee7bdbaff911794f714': 'Low Liquidity Attack',
+  '0xe49ab07879658c258d5007ac6b88428a2b88cc5cfef206222ad94690840be87a': 'Low Liquidity Attack',
   '0xb6bdedbae67cc82e60aad02a8ffab3ccbefeaa876ca7e4f291c07c798a95e339': 'Large Borrows Governance',
-  '0x916603512086fcad84c35858d2fc5356c512f72b19c80e52e8f9c04d8122e2ba': 'Community Multi-Sig',
+  '0x2e7f036f3495fec41a3eabae03b3efd378f6834bbb2976b99dfed6d3c7de7458': 'Community Multi-Sig',
   '0x0d3cdcc2757cd7837e3b302a9889c854044a80835562dc8060d7c163fbb69d53': 'Large Delegations',
-  '0xe200d890a67d51c3610520dd9fdfa9e2bd6dd341d41e32fa457601e73c4c6685': 'Oracle Price',
-  '0xf836bda7810aa2dd9df5bb7ac748f173b945863e922a15bb7c57da7b0e6dab05': 'Underlying Asset',
-  '0x125c36816fbad9974a452947bf6a98d975988ddf4342c159a986383b64765e22': 'Market Activity',
+  '0x32facccd163300ad76c7fe88b559b812dca9050a569417b42fced95594dda08e': 'Oracle Price',
+  '0xfa3044aa08927163ff8578fb5c108978dfde3a12e0b21834e53111e2859f3a59': 'Underlying Asset',
+  '0xab39733ddf86340b8e7940ebb933bb48506a341485c7f8c77da17bc1400fe028': 'Market Activity',
   '0xa0424dfee87cc34b9ff6a1dfa2cb22dbf1b20a238698ae0eeffbf07f869e5b39': 'Governance Activity',
 };
-
-const botIds = Object.keys(botIdsToNames);
 
 // override the key-value store Class
 jest.mock('defender-kvstore-client', () => ({
   KeyValueStoreClient: jest.fn().mockReturnValue(mockKeyValueStore),
 }));
 
-const { handler } = require('../datadog_forta_bot_alerts_dev/autotask-1/index');
+const { handler } = require('../datadog_alerts_heat_map/autotask-1/index');
 
 describe('Run the Autotask', () => {
   let outputObject;
@@ -76,7 +74,7 @@ describe('Run the Autotask', () => {
   it('sends alert information to Datadog when alerts are updated', async () => {
     const mockLastUpdateTimestamp = '8675308';
     const agentInformationStored = {};
-    botIds.forEach((botId) => {
+    Object.keys(botIdsToNames).forEach((botId) => {
       agentInformationStored[botId] = {};
     });
 
@@ -90,7 +88,6 @@ describe('Run the Autotask', () => {
     const mockSeverity = 'Mock Severity';
     const mockName = 'Mock Alert Name';
     const mockAlertId = 'MOCK-ALERT-ID';
-    const botId = '0x5a00b44b2db933d4c797e6bd3049abdeb89cc9ec1b2eaee7bdbaff911794f714';
     const mockDescription = 'Mock description of a Forta Bot alert';
     const alerts = [
       {
@@ -104,7 +101,7 @@ describe('Run the Autotask', () => {
         source: {
           txHash: '0xMOCKTRANSACTIONHASH',
           agent: {
-            id: botId,
+            id: '0xMOCKBOTID',
             name: null,
           },
           block: {
@@ -122,7 +119,7 @@ describe('Run the Autotask', () => {
       const {
         data: {
           query,
-          date_happened: dateHappened,
+          series,
         },
       } = inputObject;
 
@@ -130,7 +127,7 @@ describe('Run the Autotask', () => {
         return mockAlertResponse;
       }
 
-      if (dateHappened !== undefined) {
+      if (series !== undefined) {
         capturedDatadogRequest = inputObject;
       }
 
@@ -148,19 +145,28 @@ describe('Run the Autotask', () => {
         'DD-API-KEY': 'fakeApiKey',
       },
       method: 'post',
-      url: 'https://api.datadoghq.com/api/v1/events',
+      url: 'https://api.datadoghq.com/api/v2/series',
       data: {
-        date_happened: new Date(mockTimestamp).valueOf() / 1000,
-        host: 'Low Liquidity Attack',
-        tags: [
-          `botid:${botId}`,
-          `protocol:${mockProtocolName}`,
-          `severity:${mockSeverity}`,
-          'version:4',
+        series: [
+          {
+            metric: 'alertHeatMap',
+            points: [
+              {
+                timestamp: 1653150298,
+                value: 7,
+              },
+            ],
+            tags: [
+              'botid:0xMOCKBOTID',
+              `protocol:${mockProtocolName}`,
+              `severity:${mockSeverity}`,
+              'dayofweek:7',
+              'weekselapsed:23',
+              'botname:undefined',
+            ],
+            type: 0,
+          },
         ],
-        text: mockDescription,
-        aggregation_key: mockAlertId,
-        title: 'Low Liquidity Attack',
       },
     });
     expect(outputObject.lastUpdateTimestampAlerts).toStrictEqual('8675309');
